Make drag-drop pure
